backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pillow==10.2.0
tensorflow==2.15.0
numpy==1.26.4